refactored slice_plotter
