+++ models/slice/slice_algorithm.py
@@ -1,0 +1,8 @@
+
+
+class SliceAlgorithm(object):
+
+    def compute_slice(self, selected_workspace, x_axis, y_axis, smoothing):
+        """Return a numpy masked matrix ready for plotting and an error code"""
+        pass
+
